Modify a4a API to take signature as an ArrayBuffer.
