+++ extensions/amp-ad-network-adsense-impl/0.1/test/test-amp-ad-network-adsense-impl.js
@@ -15,6 +15,7 @@
  */
 
 import {AmpAdNetworkAdsenseImpl} from '../amp-ad-network-adsense-impl';
+import {base64ToByteArray} from '../../../../ads/google/a4a/utils';
 import * as sinon from 'sinon';
 
 describe('amp-ad-network-adsense-impl', () => {
@@ -60,30 +61,30 @@ describe('amp-ad-network-adsense-impl', () => {
 
   describe('#extractCreativeAndSignature', () => {
     it('without signature', () => {
-      const creativeArrayBuffer =
+      const creative =
         new TextEncoder('utf-8').encode('some creative');
       return expect(adsenseImpl.extractCreativeAndSignature(
-        creativeArrayBuffer,
+        creative,
         {
           get: function() { return undefined; },
           has: function() { return false; },
         })).to.eventually.deep.equal(
-              {creativeArrayBuffer, signature: null});
+              {creative, signature: null});
     });
     it('with signature', () => {
-      const creativeArrayBuffer =
+      const creative =
         new TextEncoder('utf-8').encode('some creative');
       return expect(adsenseImpl.extractCreativeAndSignature(
-        creativeArrayBuffer,
+        creative,
         {
           get: function(name) {
-            return name == 'X-AmpAdSignature' ? 'some_sig' : undefined;
+            return name == 'X-AmpAdSignature' ? 'AQAB' : undefined;
           },
           has: function(name) {
             return name === 'X-AmpAdSignature';
           },
         })).to.eventually.deep.equal(
-            {creativeArrayBuffer, signature: 'some_sig'});
+            {creative, signature: base64ToByteArray('AQAB')});
     });
   });
 });
